Split up program
