+++ .github/workflows/update.yml
@@ -27,7 +27,7 @@ jobs:
           python -m pip install -r requirements.txt
 
       - name: Run script
-        run: python find-datalad-repos.py
+        run: python -m find_datalad_repos
         env:
           GITHUB_TOKEN: ${{ secrets.SEARCH_GITHUB_TOKEN }}
