+++ find_datalad_repos/config.py
@@ -1,0 +1,23 @@
+RECORD_FILE = "datalad-repos.json"
+
+README_FOLDER = "READMEs"
+
+OURSELVES = {
+    "con",
+    "dandi",
+    "dandisets",
+    "datalad",
+    "datalad-collection-1",
+    "datalad-datasets",
+    "datalad-handbook",
+    "datalad-tester",
+    "dbic",
+    "jsheunis",
+    "jwodder",
+    "loj",
+    "mih",
+    "myyoda",
+    "proj-nuisance",
+    "psychoinformatics-de",
+    "yarikoptic",
+}
