+++ noxfile.py
@@ -6,11 +6,11 @@ nox.options.reuse_existing_virtualenvs = True
 @nox.session
 def run(session):
     session.install("-r", "requirements.txt")
-    session.run("find-datalad-repos.py", *session.posargs)
+    session.run("python", "-m", "find_datalad_repos", *session.posargs)
 
 
 @nox.session
 def typing(session):
     session.install("-r", "requirements.txt")
     session.install("mypy", "types-requests")
-    session.run("mypy", "find-datalad-repos.py")
+    session.run("mypy", "find_datalad_repos")
